conversational-AI/requirements.txt
openai
SQLAlchemy
PyPDF2
chromadb
langchain
python-dotenv
groq
shutilwhich
requests
fastapi
python-multipart
psutil